requirements.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
auto-py-to-exe==2.27.0
Babel==2.13.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.11.1
bleach==6.0.0
bottle==0.12.23
bottle-websocket==0.2.9
buildozer==1.5.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
coincurve==18.0.0
colorama==0.4.6
comm==0.1.2
comtypes==1.1.14
contourpy==1.0.7
customtkinter==5.2.2
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.26.3
distlib==0.3.6
Django==4.2.5
dm-tree==0.1.8
docutils==0.20.1
Eel==0.14.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.10.7
findspark==2.0.1
Flask==2.2.2
flask-ngrok==0.0.25
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.2.0
future==0.18.3
gast==0.5.4
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.17.1
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-cloud-texttospeech==2.16.1
google-pasta==0.2.0
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
greenlet==2.0.1
grpcio==1.60.1
grpcio-status==1.60.1
gTTS==2.5.1
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.21.3
hyperframe==5.2.0
idna==2.10
imageio==2.33.1
importlib-metadata==7.0.1
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.3
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==7.4.9
jupyter_core==5.1.5
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.25.0
keras==2.15.0
kiwisolver==1.4.4
lab==7.2
lazy_loader==0.3
libclang==16.0.6
lmdb==1.4.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.2.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
mysql-connector-python==8.0.32
namex==0.0.7
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
openai==0.26.5
opencv-python==4.7.0.68
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.4.0
packaging==23.0
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pefile==2022.5.30
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.11.0
playsound==1.2.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.23.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.53
pycaw==20220416
pycoingecko==3.1.0
pycparser==2.21
pygame==2.3.0
PyGetWindow==0.0.9
Pygments==2.14.0
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2022.15
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
pyspark==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
pyttsx3==2.90
pytweening==1.0.4
pytz==2023.3.post1
pywhatkit==3.8
pywin32==305
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.2
screen-brightness-control==0.18.0
selenium==4.9.0
Send2Trash==1.8.2
sh==2.0.3
simplejson==3.18.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
sqlparse==0.4.4
stack-data==0.6.2
statistics==1.0.3.5
stats==0.1.2a0
swig==4.1.1.post1
sympy==1.12
tb-nightly==2.16.0a20240107
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
tf_keras-nightly==2.16.0.dev2024010710
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.38.2
trio==0.22.0
trio-websocket==0.10.2
txt2tags==3.7
typing_extensions==4.4.0
tzdata==2023.3
ufal.pybox2d==2.3.10.2
uri-template==1.2.0
urllib3==1.26.14
virtualenv==20.21.0
waitress==2.1.2
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
whichcraft==0.6.1
widgetsnbextension==4.0.5
wikipedia==1.4.0
WMI==1.5.1
wrapt==1.14.1
wsproto==1.2.0
yapf==0.40.2
yarl==1.8.2
zipp==3.17.0
zope.event==4.6
zope.interface==5.5.2
